added logic to check of remote users

+++ manifests/stig/v79067.pp
@@ -4,32 +4,37 @@
 # SQL Server must protect against a user falsely repudiating by ensuring
 # only clearly unique Active Directory user accounts can connect to the database.
 #
-# Not checking for the output of this command to determine if account is local:
-# ([ADSISearcher]"(&(!ObjectCategory=Computer)(Name=<name>))").FindAll()
-#
-# Assume the '$' at the end means its shared so lets remove it.
-#
-#
-# U    WINDOWS_LOGIN
-# G    GROUP
+# NOTE: Not checking for the output of this command to determine if account is local:
+#       ([ADSISearcher]"(&(!ObjectCategory=Computer)(Name=<name>))").FindAll()
+# NOTE:  USE fact "sqlserver_shared_database_accounts_detail", it includes
+#        the powershell check code, but we don't know how the output is returned yet.
 #
 define secure_sqlserver::stig::v79067 (
   Boolean       $enforced = false,
   String[1,16]  $instance = 'MSSQLSERVER',
   String        $database,
 ) {
+
   if $enforced {
+
+    # yaml file contains approved_users, skip the DROP for any in the list...
+    $approved_users = lookup('secure_sqlserver::approved_shared_accounts')
     $shared_accounts = $facts['sqlserver_shared_database_accounts']
+
     unless $shared_accounts == undef or empty($shared_accounts) {
       $shared_accounts.each |$drop_user| {
-        $sql = "DROP USER IF EXISTS ${drop_user}"
-        ::secure_sqlserver::log { "v79067 sql = \n${sql}": }
-        sqlserver_tsql{ "v79067_drop_shared_user_${database}_${username}":
-          instance => $instance,
-          command  => $sql,
-          require  => Sqlserver::Config[$instance],
+        unless $drop_user in $approved_users {
+          $sql = "DROP USER IF EXISTS ${drop_user}"
+          ::secure_sqlserver::log { "v79067 sql = \n${sql}": }
+          sqlserver_tsql{ "v79067_drop_shared_user_${database}_${username}":
+            instance => $instance,
+            command  => $sql,
+            require  => Sqlserver::Config[$instance],
+          }
         }
       }
     }
+
   }
+
 }
